misc. bugfix

+++ src/platform/clock_realtime.hpp
@@ -7,6 +7,6 @@
 namespace Hamster
 {
     // Offset from the system clock
-    extern int32_t clock_rt_offset;
+    extern int64_t clock_rt_offset;
 } // namespace Hamster
